Rule reorganization
- Added new section in F
- Rules needed rearranging to match KLC
