+++ pcb/rules/F7_6.py
