+++ pcb/rules/__init__.py
@@ -4,10 +4,6 @@ __all__ = [
 "G1_1",
 "G1_7",
 
-"F4_1",
-"F4_2",
-"F4_3",
-
 "F5_1",
 "F5_2",
 "F5_3",
@@ -15,12 +11,16 @@ __all__ = [
 "F6_1",
 "F6_2",
 "F6_3",
-"F6_4",
-"F6_5",
-"F6_6",
 
-"F8_1",
-"F8_2",
-"F8_3",
+"F7_1",
+"F7_2",
+"F7_3",
+"F7_4",
+"F7_5",
+"F7_6",
+
+"F9_1",
+"F9_2",
+"F9_3",
 
 ]
